@@ -21,8 +21,8 @@ type Noop struct {
 }
 
 func (n Noop) Run(_ context.Context, lintCtx *Context) ([]result.Issue, error) {
-	lintCtx.Log.Warnf("%s is disabled because of go1.18."+
-		" You can track the evolution of the go1.18 support by following the https://github.com/golangci/golangci-lint/issues/2649.", n.name)
+	lintCtx.Log.Warnf("%s is disabled because of generics."+
+		" You can track the evolution of the generics support by following the https://github.com/golangci/golangci-lint/issues/2649.", n.name)
 	return nil, nil
 }
 
